Refactor file structure and update dependencies
