+++ src/components/Header.tsx
@@ -1,10 +1,17 @@
 import React from 'react';
 import { ConfigProvider, Menu } from 'antd';
-import { Link, useLocation } from 'react-router-dom';
+import Link from 'next/link';
+import { useRouter } from 'next/router';
 import { GithubOutlined } from '@ant-design/icons';
 import useScrollPosition from '../utils/useScrollPosition';
 
-const menuItems = [
+interface MenuItem {
+  key: string;
+  link: string;
+  label: string;
+}
+
+const menuItems: MenuItem[] = [
   { key: '1', link: '/', label: 'Home' },
   { key: '2', link: '/team', label: 'Team' },
   { key: '3', link: '/resources', label: 'Resources' },
@@ -14,9 +21,9 @@ const menuItems = [
   { key: '7', link: '/contact', label: 'Contact' },
 ];
 
-const Header = () => {
-  const location = useLocation();
-  const route = location.pathname;
+const Header: React.FC = () => {
+  const router = useRouter();
+  const route = router.pathname;
   const { scrolled100vh } = useScrollPosition();
   const tintedHeader = scrolled100vh || route !== "/";
   if (route === "/test") return null;
@@ -30,13 +37,15 @@ const Header = () => {
           tintedHeader ? 'bg-white shadow-lg' : 'bg-transparent'
         } pr-4 flex items-center h-16 w-full top-0 left-0 z-50 transition-all duration-300 ease-in-out`}
       >
-        <a href="/" className="flex items-center">
-          <img
-            src="/assets/gitamlogo.png"
-            alt="Gitam Logo"
-            className="h-10 ml-4 transform transition-transform duration-500 hover:scale-110"
-          />
-        </a>
+        <Link href="/" passHref>
+          <a className="flex items-center">
+            <img
+              src="/assets/gitamlogo.png"
+              alt="Gitam Logo"
+              className="h-10 ml-4 transform transition-transform duration-500 hover:scale-110"
+            />
+          </a>
+        </Link>
         <Menu
           theme="light"
           mode="horizontal"
@@ -45,14 +54,14 @@ const Header = () => {
         >
           {menuItems.map((item) => (
             <Menu.Item key={item.key} className="hover:bg-gray-200 rounded-lg">
-              <Link to={item.link}>
-                <span
+              <Link href={item.link} passHref>
+                <a
                   className={`${
                     tintedHeader ? 'text-black' : 'text-white'
                   } font-semibold`}
                 >
                   {item.label}
-                </span>
+                </a>
               </Link>
             </Menu.Item>
           ))}
