+++ src/components/Navbar.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Menu } from 'antd';
-import { Link } from 'react-router-dom';
+import Link from 'next/link';
 
-function Navbar() {
+const Navbar: React.FC = () => {
   return (
     <nav style={{ backgroundColor: '#001529', padding: '0 50px' }}>
       <Menu
@@ -11,26 +11,38 @@ function Navbar() {
         style={{ display: 'flex', justifyContent: 'center', backgroundColor: 'transparent', borderBottom: 'none' }}
       >
         <Menu.Item key="1">
-          <Link to="/">Home</Link>
+          <Link href="/" passHref>
+            <a>Home</a>
+          </Link>
         </Menu.Item>
         <Menu.Item key="2">
-          <Link to="/resources">Resources</Link>
+          <Link href="/resources" passHref>
+            <a>Resources</a>
+          </Link>
         </Menu.Item>
         <Menu.Item key="3">
-          <Link to="/team">Team</Link>
+          <Link href="/team" passHref>
+            <a>Team</a>
+          </Link>
         </Menu.Item>
         <Menu.Item key="4">
-          <Link to="/gallery">Gallery</Link>
+          <Link href="/gallery" passHref>
+            <a>Gallery</a>
+          </Link>
         </Menu.Item>
         <Menu.Item key="5">
-          <Link to="/projects">Projects</Link>
+          <Link href="/projects" passHref>
+            <a>Projects</a>
+          </Link>
         </Menu.Item>
         <Menu.Item key="6">
-          <Link to="/testimonials">Testimonials</Link>
+          <Link href="/testimonials" passHref>
+            <a>Testimonials</a>
+          </Link>
         </Menu.Item>
       </Menu>
     </nav>
   );
-}
+};
 
 export default Navbar;
