--- src/pages/TestimonialsPage/TestimonialsPage.css
@@ -1,81 +1,0 @@
-/* Modal Glass Effect */
-.modal-glass-effect .ant-modal-content {
-  background: rgba(255, 255, 255, 0.2); 
-  -webkit-backdrop-filter: blur(15px) saturate(180%); 
-  backdrop-filter: blur(15px) saturate(180%); 
-  border-radius: 20px; 
-  box-shadow: 0 8px 32px rgba(31, 38, 135, 0.37); 
-  border: 1px solid rgba(255, 255, 255, 0.18); 
-}
-
-/* Zoom Animation */
-.zoom-enter {
-  opacity: 0;
-  transform: scale(0.95);
-}
-
-.zoom-enter-active {
-  opacity: 1;
-  transform: scale(1);
-  transition: opacity 0.3s ease, transform 0.3s ease;
-}
-
-.zoom-leave {
-  opacity: 1;
-  transform: scale(1);
-}
-
-.zoom-leave-active {
-  opacity: 0;
-  transform: scale(0.95);
-  transition: opacity 0.3s ease, transform 0.3s ease;
-}
-
-/* Confetti Container */
-.confetti-container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  pointer-events: none;
-  z-index: 9999; 
-}
-
-/* Heading Styling */
-.heading-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 1rem; /* Adjust the bottom margin as needed */
-}
-
-.heading-effect {
-  position: relative;
-  font-size: 2.5rem;
-  font-weight: bold;
-  text-align: center;
-  padding: 0.5rem 1rem;
-  background: linear-gradient(90deg, #1890ff 30%, #cdb90c 90%);
-  -webkit-background-clip: text;
-  background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-
-.heading-effect::before, .heading-effect::after {
-  content: '✨';
-  font-size: 1.5rem;
-  margin: 0 0.5rem;
-  color: #485ec0;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-}
-
-.heading-effect::before {
-  left: -2rem;
-}
-
-.heading-effect::after {
-  right: -2rem;
-}
